Add files via upload
This winrar represents my first ever project that I created alone. With the use of Java Swing, I created this game called Wheel of Fortune. Also, I used the application called NetBeans 12.0. The rules of this specific game are the following:
1. There are 3 players on this game
2. There are 10 rounds on this game
3. There are 10 categories of 10 secret phrases(on this game the categories are these in order: 1.
4. When the category is written, you spin the wheel
5. After that you get a phrase that gives a clue about the secret phrase
6. If you guess the whole phrase correctly, you get the amount of money you got on the wheel and you continue to the next round, else you lose and the next player spins the wheel and tries to guess the secret phrase
7. The winner of this game is the one with the most earned money

Please, be free to try running this game and give me your honest opinion and suggestions about how to make it more advanced